requirements.txt
google-cloud-storage
google-cloud-aiplatform
pandas
fastapi
uvicorn[standard]
python-multipart
